refactor for background and standard
