+++ src/cosmology/compat/astropy/__init__.py
@@ -14,13 +14,15 @@ There are the following required objects for a Cosmology-API compatible library:
 """
 
 from cosmology.compat.astropy import constants
+from cosmology.compat.astropy.background import AstropyBackgroundCosmology
 from cosmology.compat.astropy.core import AstropyCosmology
-from cosmology.compat.astropy.flrw import AstropyFLRW
+from cosmology.compat.astropy.standard import AstropyStandardCosmology
 
 __all__ = [
     # Cosmology API
     "constants",
     # Wrappers
     "AstropyCosmology",
-    "AstropyFLRW",
+    "AstropyBackgroundCosmology",
+    "AstropyStandardCosmology",
 ]
