+++ src/cosmology/compat/astropy/core.py
@@ -3,17 +3,29 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import cast
-
-from cosmology.api import CosmologyAPINamespace, CosmologyWrapperAPI
+from typing import Any, cast
 
 import astropy.cosmology as astropy_cosmology
+from numpy import floating
+from numpy.typing import NDArray
+from typing_extensions import TypeAlias
+
+from cosmology.api import CosmologyAPINamespace, CosmologyWrapperAPI
 
 __all__: list[str] = []
 
 
+################################################################################
+# PARAMETERS
+
+NDFloating: TypeAlias = NDArray[floating[Any]]
+
+
+################################################################################
+
+
 @dataclass(frozen=True)
-class AstropyCosmology(CosmologyWrapperAPI):
+class AstropyCosmology(CosmologyWrapperAPI[NDFloating]):
     """The Cosmology API wrapper for :mod:`astropy.cosmology.Cosmology`."""
 
     cosmo: astropy_cosmology.Cosmology
